requirements.txt
# Core dependencies
numpy>=1.19.0
scipy>=1.5.0
scikit-image>=0.17.0
tqdm>=4.50.0
opencv-python>=4.5.0
matplotlib>=3.3.0
pandas>=1.2.0
seaborn>=0.11.0
pyyaml>=5.4.0

# Optional but recommended for advanced features
plotly>=5.3.0

# Optional for 3D visualization (difficult to install, use conda if possible)
# mayavi>=4.7.0

# For Excel export support
openpyxl>=3.0.0
